backend/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==4.0.4
django-filter==21.1
django-picklefield==3.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
idna==3.3
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.1
pytz==2022.1
requests==2.27.1
sqlparse==0.2.4
tzdata==2022.1
urllib3==1.26.9
